iota_sdk_graphql_client::error

Type Alias QueryError

pub type QueryError = GraphQlError<ErrorExtensions>;

A GraphQL error as reported by the server, with its ErrorExtensions decoded.

Aliased Type

pub struct QueryError {
    pub message: String,
    pub locations: Option<Vec<GraphQlErrorLocation>>,
    pub path: Option<Vec<GraphQlErrorPathSegment>>,
    pub extensions: Option<ErrorExtensions>,
}

Fields
message: String

A description of the error which has taken place.

locations: Option<Vec<GraphQlErrorLocation>>

Optional description of the locations where the errors have taken place.

path: Option<Vec<GraphQlErrorPathSegment>>

Optional path to the response field which experienced the associated error.

extensions: Option<ErrorExtensions>

Optional arbitrary extra data describing the error in more detail.