iota_sdk_graphql_client

Module query_types

Modules

schema

Structs

Base64

BigInt

CoinMetadata

The coin metadata associated with the given coin type.

DateTime

Epoch

Event

EventFilter

GraphQLAddress

MoveData

MoveModuleQueryFragment

MoveObject

MovePackageAddress

MoveType

MoveViewResult

The result of executing a Move View Function.

ObjectFilter

ObjectRef

PageInfo

Information about pagination in a connection.

ProtocolConfigAttr

A key-value protocol configuration attribute.

ProtocolConfigFeatureFlag

Feature flags are a form of boolean configuration that are usually used to gate features while they are in development. Once a lag has been enabled, it is rare for it to be disabled.

ProtocolConfigs

Information about the configuration of the protocol. Constants that control how the chain operates. These can only change during protocol upgrades which happen on epoch boundaries.

ServiceConfig

TransactionMetadata

TransactionsFilter

Filter for transaction queries.

Validator

Represents a validator in the system.

ValidatorCredentials

The credentials related fields associated with a validator.

ValidatorSet

Enums

AddressTransactionRelationship

The relationship between an address and a transaction.

Feature

SubscriptionEventFilter

Filter incoming events in a subscription.

SubscriptionTransactionFilter

Filter incoming transactions in a subscription.

TransactionBlockKindInput

TransactionsSelector

Selection criteria for querying transactions.