iota_sdk_graphql_client

Module output_types

Structs

BcsName

Helper struct for passing a raw bcs value.

DryRunEffect

Effects of a single command in the dry run, including mutated references and return values.

DryRunMutation

A mutation to an argument that was mutably borrowed by a command.

DryRunResult

The result of a simulation (dry run), which includes the effects of the transaction and intermediate results for each command.

DryRunReturn

A return value from a command in the dry run.

DynamicFieldName

The name part of a dynamic field, including its type, bcs, and json representation.

DynamicFieldOutput

The output of a dynamic field query, that includes the name, value, and value’s json representation.

DynamicFieldValue

The value part of a dynamic field.

NameValue

Helper struct for passing a value that has a type that implements Serialize, for the dynamic fields API.

TransactionDataEffects

Enums

TransactionArgument

A transaction argument used in programmable transactions.