iota_sdk_graphql_client

Module query_types

Modules

schema

Structs

ActiveValidatorsArgs

ActiveValidatorsQueryFragment

AddressTransactionBlocksQueryFragment

AddressTransactionsQueryArgs

AddressTransactionsQueryFragment

Balance

BalanceArgs

BalanceQueryFragment

Base64

BigInt

ChainIdentifierQueryFragment

CheckpointArgs

CheckpointId

CheckpointQueryFragment

CheckpointTotalTxQueryFragment

CheckpointsArgs

CheckpointsQueryFragment

CoinMetadata

The coin metadata associated with the given coin type.

CoinMetadataArgs

CoinMetadataQueryFragment

DateTime

DryRunArgs

DryRunEffect

DryRunMutation

DryRunQueryFragment

DryRunResult

DryRunReturn

DynamicFieldArgs

DynamicFieldConnectionArgs

DynamicFieldName

DynamicFieldQueryFragment

DynamicFieldsOwnerQueryFragment

DynamicObjectFieldQueryFragment

Epoch

EpochArgs

EpochQueryFragment

EpochSummaryQueryFragment

EpochValidator

Event

EventConnection

EventFilter

EventsQueryArgs

EventsQueryFragment

EventsSubscription

EventsSubscriptionArgs

ExecuteTransactionArgs

ExecuteTransactionQueryFragment

ExecutionResult

GasCoin

GraphQLAddress

Input

IotaNamesAddressDefaultNameQueryFragment

IotaNamesAddressRegistrationsQueryFragment

IotaNamesDefaultNameArgs

IotaNamesDefaultNameQueryFragment

IotaNamesRegistrationsArgs

IotaNamesRegistrationsQueryFragment

Lagged

Emitted when the server dropped one or more payloads before this one because the subscriber could not keep up.

LatestPackageQueryFragment

MoveData

MoveEnum

MoveEnumConnection

MoveEnumVariant

MoveField

MoveFunction

MoveFunctionConnection

MoveFunctionTypeParameter

MoveModule

MoveModuleConnection

MoveModuleQueryFragment

MoveObject

MoveObjectContents

MovePackageConnection

MovePackageQueryFragment

MovePackageVersionFilter

MoveStructConnection

MoveStructQueryFragment

MoveStructTypeParameter

MoveType

MoveValue

MoveViewCallArgs

Arguments for the Move View Call GraphQL query.

MoveViewCallQueryFragment

GraphQL query for executing a Move View Function.

MoveViewResult

The result of executing a Move View Function.

NameRegistration

NameRegistrationConnection

NormalizedMoveFunctionQueryArgs

NormalizedMoveFunctionQueryFragment

NormalizedMoveModuleQueryArgs

NormalizedMoveModuleQueryFragment

ObjectFilter

ObjectQueryArgs

ObjectQueryFragment

ObjectRef

ObjectsQueryArgs

ObjectsQueryFragment

OpenMoveType

Owner

PackageArgs

PackageCheckpointFilter

PackageQueryFragment

PackageVersionsArgs

PackageVersionsQueryFragment

PackagesQueryArgs

PackagesQueryFragment

PageInfo

Information about pagination in a connection.

ProtocolConfigAttr

A key-value protocol configuration attribute.

ProtocolConfigFeatureFlag

Feature flags are a form of boolean configuration that are usually used to gate features while they are in development. Once a lag has been enabled, it is rare for it to be disabled.

ProtocolConfigQueryFragment

ProtocolConfigs

Information about the configuration of the protocol. Constants that control how the chain operates. These can only change during protocol upgrades which happen on epoch boundaries.

ProtocolVersionArgs

ResolveIotaNamesAddressArgs

ResolveIotaNamesAddressQueryFragment

ResultArg

ServiceConfig

ServiceConfigQueryFragment

SubscriptionEvent

An event as delivered over a subscription. Mirrors Event but also selects the emitting transaction’s digest, which is used as the stream recovery cursor (startAfter).

SubscriptionEventFilter

Filter incoming events in a subscription. Exactly one field must be set (the GraphQL input is @oneOf).

SubscriptionTransactionBlock

A transaction block as delivered over a subscription. Selects the digest (used as the stream recovery cursor) alongside the SenderSignedData BCS that a SignedTransaction is rebuilt from.

SubscriptionTransactionFilter

Filter incoming transactions in a subscription. Exactly one field must be set (the GraphQL input is @oneOf).

TransactionBlock

TransactionBlockArgs

TransactionBlockCheckpointQueryFragment

TransactionBlockEffectsQueryFragment

TransactionBlockFilter

The GraphQL input object, built from a TransactionsFilter.

TransactionBlockIndexedQueryFragment

TransactionBlockQueryFragment

TransactionBlockWithEffects

TransactionBlockWithEffectsQueryFragment

TransactionBlocksEffectsQueryFragment

TransactionBlocksQueryArgs

TransactionBlocksQueryFragment

TransactionBlocksWithEffectsQueryFragment

TransactionMetadata

TransactionsFilter

Filter for transaction queries.

TransactionsSubscription

TransactionsSubscriptionArgs

Validator

Represents a validator in the system.

ValidatorConnection

ValidatorCredentials

The credentials related fields associated with a validator.

ValidatorSet

ValidatorSetQueryFragment

Enums

AddressTransactionRelationship

The relationship between an address and a transaction.

EventSubscriptionPayload

Feature

MoveAbility

MoveVisibility

TransactionArgument

TransactionBlockKindInput

TransactionBlockSubscriptionPayload

TransactionsSelector

Selection criteria for querying transactions.